infra/pyproject.toml
[project]
name = "nutriwise-infra"
version = "0.1.0"
description = "NutriWise AWS CDK infrastructure"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.160.0",
    "constructs>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nutriwise_cdk*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
